LPD-59495 Rename utils to camelCase
