+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/categorization/tags/EditTagsModal.tsx
@@ -16,7 +16,7 @@ import ApiHelper from '../../../common/services/ApiHelper';
 import {
 	displayErrorToast,
 	displayNameInUseErrorToast,
-} from '../../../common/utils/ToastUtil';
+} from '../../../common/utils/toastUtil';
 import CategorizationSpaces from '../components/CategorizationSpaces';
 
 export default function EditTagsModalContent({
