+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/categorization/vocabularies/EditVocabulary.tsx
@@ -18,7 +18,7 @@ import {IVocabulary} from '../../../common/types/IVocabulary';
 import {
 	displayNameInUseErrorToast,
 	displaySystemErrorToast,
-} from '../../../common/utils/ToastUtil';
+} from '../../../common/utils/toastUtil';
 import {DEFAULT_PERMISSIONS} from '../utils/CategorizationPermissionsUtil';
 import ConfirmChangesModal from './ConfirmChangesModal';
 import EditAssociatedAssetTypes from './EditAssociatedAssetTypes';
