+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main_view/props_transformer/VocabularyFDSPropsTransformer.ts
@@ -6,7 +6,7 @@
 import {IInternalRenderer} from '@liferay/frontend-data-set-web';
 
 import {IVocabulary} from '../../common/types/IVocabulary';
-import {openGenericFDSDeleteConfirmationModal} from '../../common/utils/GenericOpenModalUtil';
+import {openGenericFDSDeleteConfirmationModal} from '../../common/utils/genericOpenModalUtil';
 import MultipleSpacesRenderer from './cell_renderers/MultipleSpacesRenderer';
 import VocabularyRenderer from './cell_renderers/VocabularyRenderer';
